% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duane_plot.R
\name{duane_plot}
\alias{duane_plot}
\title{Duane Plot.}
\usage{
duane_plot(times, failures)
}
\arguments{
\item{times}{A vector of cumulative times at which failures occurred.}

\item{failures}{A vector of the number of failures at each corresponding time in times.}
}
\value{
The function does not return a value.
}
\description{
Duane Plot.
}
\examples{
times <- c(100, 200, 300, 400, 500)
failures <- c(1, 2, 1, 3, 2)
fit <- duane_plot(times, failures)
summary(fit)
}
