% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grwr_api.R
\name{grwr_api}
\alias{grwr_api}
\title{ReliaGrowR API}
\usage{
grwr_api()
}
\value{
Launches the ReliaGrowR API on a local server.
}
\description{
This function provides an interface to the ReliaGrowR API.#' This function provides an interface to the ReliaGrowR API.
}
\examples{
\dontrun{
grwr_api()
}
}
