% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duane.R
\name{print.duane_predict}
\alias{print.duane_predict}
\title{Print Method for duane_predict Objects}
\usage{
\method{print}{duane_predict}(x, ...)
}
\arguments{
\item{x}{An object of class \code{duane_predict}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
Prints a formatted table of forecast MTBF with confidence bounds.
}
\examples{
times <- c(100, 200, 300, 400, 500)
failures <- c(1, 2, 1, 3, 2)
fit <- duane(times, failures)
fc <- predict_duane(fit, times = c(1000, 2000))
print(fc)
}
\seealso{
Other Duane functions: 
\code{\link{duane}()},
\code{\link{plot.duane}()},
\code{\link{plot.duane_predict}()},
\code{\link{predict_duane}()},
\code{\link{print.duane}()}
}
\concept{Duane functions}
