% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{print.gof}
\alias{print.gof}
\title{Print Method for gof Objects}
\usage{
\method{print}{gof}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gof}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
Prints a summary of goodness-of-fit statistics.
}
\examples{
times <- c(5, 10, 15, 20, 25)
failures <- c(1, 2, 1, 3, 2)
fit <- rga(times, failures)
g <- gof(fit)
print(g)
}
\seealso{
Other goodness-of-fit: 
\code{\link{gof}()},
\code{\link{ppplot.rga}()},
\code{\link{qqplot.rga}()}
}
\concept{goodness-of-fit}
