% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp.R
\name{print.nhpp}
\alias{print.nhpp}
\title{Print Method for nhpp Objects.}
\usage{
\method{print}{nhpp}(x, ...)
}
\arguments{
\item{x}{An object of class \code{nhpp}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
Prints a summary of the NHPP model results.
}
\examples{
time <- c(200, 400, 600, 800, 1000)
event <- c(3, 5, 4, 7, 6)
result <- nhpp(time, event)
print(result)
}
\seealso{
Other Repairable Systems Analysis: 
\code{\link{exposure}()},
\code{\link{mcf}()},
\code{\link{nhpp}()},
\code{\link{plot.exposure}()},
\code{\link{plot.mcf}()},
\code{\link{plot.nhpp}()},
\code{\link{plot.nhpp_predict}()},
\code{\link{predict_nhpp}()},
\code{\link{print.exposure}()},
\code{\link{print.mcf}()},
\code{\link{print.nhpp_predict}()}
}
\concept{Repairable Systems Analysis}
