% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdt.R
\name{print.rdt}
\alias{print.rdt}
\title{Print method for rdt objects}
\usage{
\method{print}{rdt}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rdt}.}

\item{...}{Additional arguments (not used).}
}
\description{
This function provides a formatted print method for objects of class \code{rdt}.
}
