% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rga.R
\name{plot_rga}
\alias{plot_rga}
\title{Plotting Function for Reliability Growth Analysis.}
\usage{
plot_rga(times, failures, result, conf_level = 0.95)
}
\arguments{
\item{times}{A vector of cumulative times at which failures occurred.}

\item{failures}{A vector of the number of failures at each corresponding time in times.}

\item{result}{A list of results for a Reliability Growth Analysis.}

\item{conf_level}{the desired confidence level, which defaults to 95\%.}
}
\value{
The function does not return a value.
}
\description{
Plotting Function for Reliability Growth Analysis.
}
\examples{
times <- c(100, 200, 300, 400, 500)
failures <- c(1, 2, 1, 3, 2)
result <- rga(times, failures)
plot_rga(times, failures, result)
}
