% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{ppplot.rga}
\alias{ppplot.rga}
\title{P-P Plot for RGA Objects}
\usage{
ppplot.rga(x, main = "P-P Plot", ...)
}
\arguments{
\item{x}{An object of class \code{rga}.}

\item{main}{Title of the plot.}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\description{
This function creates a P-P plot for a fitted Reliability Growth Analysis (RGA) model
Currently only supports the Crow-AMSAA model.
}
\seealso{
Other goodness-of-fit: 
\code{\link{qqplot.rga}()}
}
\concept{goodness-of-fit}
