% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duane.R
\name{print.duane}
\alias{print.duane}
\title{Print method for duane objects.}
\usage{
\method{print}{duane}(x, ...)
}
\arguments{
\item{x}{An object of class "duane" returned by the duane_plot function.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
This function prints a summary of the Duane analysis result.
}
\examples{
times <- c(100, 200, 300, 400, 500)
failures <- c(1, 2, 1, 3, 2)
fit <- duane(times, failures)
print(fit)
}
\seealso{
Other Duane functions: 
\code{\link{duane}()},
\code{\link{plot.duane}()}
}
\concept{Duane functions}
