% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duane_plot.R
\name{print.duane}
\alias{print.duane}
\title{Print method for duane objects.}
\usage{
\method{print}{duane}(x, ...)
}
\arguments{
\item{x}{An object of class "duane" returned by the duane_plot function.}

\item{...}{Additional arguments (not used).}
}
\description{
This function prints a summary of the Duane analysis result.
}
