% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rga.R
\name{print.rga}
\alias{print.rga}
\title{Print method for rga objects.}
\usage{
\method{print}{rga}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rga}, which contains the results from the RGA model.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
This function prints a summary of the results from an object of class \code{rga}.
}
\examples{
times <- c(100, 200, 300, 400, 500)
failures <- c(1, 2, 1, 3, 2)
result <- rga(times, failures)
print(result)
}
\seealso{
Other Reliability Growth Analysis: 
\code{\link{plot.rga}()},
\code{\link{rga}()}
}
\concept{Reliability Growth Analysis}
