% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rga.R
\name{print.rga}
\alias{print.rga}
\title{Print method for rga objects.}
\usage{
\method{print}{rga}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rga}, which contains the results from the RGA model.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for rga objects.
}
\seealso{
Other Reliability Growth Analysis: 
\code{\link{plot.rga}()},
\code{\link{rga}()}
}
\concept{Reliability Growth Analysis}
