% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rga.R
\name{print.rga}
\alias{print.rga}
\title{Print method for rga objects.}
\usage{
\method{print}{rga}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rga}.}

\item{...}{Additional arguments (not used).}
}
\description{
This function prints a summary of the RGA analysis result.
}
