% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliaGrowR-package.R
\docType{package}
\name{ReliaGrowR-package}
\alias{ReliaGrowR}
\alias{ReliaGrowR-package}
\title{ReliaGrowR: Reliability Growth Analysis}
\description{
Modeling and plotting functions for Reliability Growth Analysis (RGA). Models include the Duane (1962) \doi{10.1109/TA.1964.4319640}, Non-Homogeneous Poisson Process (NHPP) by Crow (1975) (No. AMSAATR138), Piecewise Weibull NHPP by Guo et al. (2010) \doi{10.1109/RAMS.2010.5448029}, and Piecewise Weibull NHPP with Change Point Detection based on the 'segmented' package by Muggeo (2024) \url{https://cran.r-project.org/package=segmented}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paulgovan.github.io/ReliaGrowR/}
  \item \url{https://github.com/paulgovan/ReliaGrowR}
  \item Report bugs at \url{https://github.com/paulgovan/ReliaGrowR/issues}
}

}
\author{
\strong{Maintainer}: Paul Govan \email{paul.govan2@gmail.com} (\href{https://orcid.org/0000-0002-1821-8492}{ORCID}) [copyright holder]

}
\keyword{internal}
