% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rga.R
\name{rga}
\alias{rga}
\title{Reliability Growth Analysis.}
\usage{
rga(
  times,
  failures,
  model_type = "Crow-AMSAA",
  breaks = NULL,
  conf_level = 0.95
)
}
\arguments{
\item{times}{A vector of cumulative failure times.}

\item{failures}{A vector of the number of failures at each corresponding time in times.}

\item{model_type}{The model type. Either \code{Crow-AMSAA} (default) or \verb{Piecewise NHPP} with change point detection.}

\item{breaks}{An optional vector of breakpoints for the \verb{Piecewise NHPP} model.}

\item{conf_level}{The desired confidence level, which defaults to 95\%.}
}
\value{
The function returns an object of class \code{rga} that contains:
\item{model}{The fitted model object (lm or segmented).}
\item{logLik}{The log-likelihood of the fitted model.}
\item{AIC}{Akaike Information Criterion.}
\item{BIC}{Bayesian Information Criterion.}
\item{breakpoints}{Breakpoints (log scale) if applicable.}
\item{fitted_values}{Fitted cumulative failures on the original scale.}
\item{lower_bounds}{Lower confidence bounds (original scale).}
\item{upper_bounds}{Upper confidence bounds (original scale).}
\item{betas}{Estimated beta(s).}
\item{lambdas}{Estimated lambda(s).}
}
\description{
This function performs reliability growth analysis using the Crow-AMSAA model by
Crow (1975) \url{https://apps.dtic.mil/sti/citations/ADA020296} or piecewise
NHPP model by Guo et al. (2010) \url{doi:10.1109/RAMS.2010.5448029}.
}
\examples{
times <- c(100, 200, 300, 400, 500)
failures <- c(1, 2, 1, 3, 2)
result <- rga(times, failures)
print(result)

}
\seealso{
Other Reliability Growth Analysis: 
\code{\link{plot.rga}()},
\code{\link{print.rga}()}
}
\concept{Reliability Growth Analysis}
